Declutter potentially useful but unused scripts
