+++ crustal_sz_shared_scripts/old_or_unused_scripts/check_rates.py
